requirements.txt
# --- Основные библиотеки проекта ---
python-telegram-bot==20.7
qobuz-dl==0.9.9.10
pydub
python-dotenv

# --- Вспомогательные библиотеки (pip подберет совместимые версии) ---
anyio
beautifulsoup4
certifi
charset-normalizer
colorama
h11
httpcore
httpx
idna
mutagen
pathvalidate
pick
requests
sniffio
soupsieve
tqdm
typing_extensions
urllib3
audD